services/auth/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
asyncpg
pydantic
python-jose[cryptography]
passlib[bcrypt]
redis
python-dotenv
PyJWT